+++ src/prefect/agent/kubernetes/__init__.py
@@ -1,0 +1,2 @@
+from prefect.agent.kubernetes.agent import KubernetesAgent
+from prefect.agent.kubernetes.resource_manager import ResourceManager